+++ R/plantuml_knit_engine.R
@@ -5,9 +5,10 @@
 #'    For formats which return images, these will be inserted,
 #'    formats resulting in text qill be included as text.
 #'    At the moment, the following are supported:
-#'      - **png**	To generate image using PNG format (default).
-#'      - **svg**	To generate image using SVG format.
-#'      - **eps**	To generate text in EPS format; or generates an image when
+#'      - **auto** (default) uses `svg` if output is `html` and `pdf` if output is `pdf`
+#'      - **png**	 To generate image using PNG format (default).
+#'      - **svg**	 To generate image using SVG format.
+#'      - **eps**	 To generate text in EPS format; or generates an image when
 #'      outputting LaTeX rather than HTML formats.
 #'  - **plantuml.path**: the path where the resulting files will be saved.
 #'    Default is the same directory as the `.Rmd` file is in. The path will be created if it does not exist.
@@ -38,6 +39,9 @@ plantuml_knit_engine <-  function(options) {
     path <-  "."
   } else {
     path <- options$plantuml.path
+  }
+  if (is.null(options$plantuml.format)) {
+    options$plantuml.format <-  "auto"
   }
   ###
   result <- list(out = "", code = "")