+++ tests/testthat/test-01.server.R
@@ -1,3 +1,17 @@
+
+server_stop()
+
+test_that(
+  "Server ststus before tests",
+  {
+    expect_snapshot(
+      {
+        server_status()
+      }
+    )
+  }
+)
+
 test_that(
   "Server is not starting when server_url != \"localhost\"",
   {